Update CodeSee Maps to the latest version
Heya! Thanks for taking the time to look at this.

After talking to the team, it turns out we needed to update the github action to be able to show maps when first-time contributors open a PR.

This should hopefully resolve that.

Please feel free to hit me up if you have any questions!

Thanks in advance!